Fine calculation working
